Extract live keychain to separate library
